Move `ObjectChange` to extras
