+++ extras/filters.py
@@ -1,4 +1,5 @@
 import django_filters
+from django.contrib.auth.models import User
 from django.db.models import Q
 
 from utils.filters import (
@@ -8,12 +9,13 @@ from utils.filters import (
     NameSlugSearchFilterSet,
 )
 
-from .enums import HttpMethod
+from .enums import HttpMethod, ObjectChangeAction
 from .models import (
     IXAPI,
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
+    ObjectChange,
     Tag,
     Webhook,
 )
@@ -74,6 +76,43 @@ class IXAPIFilterSet(BaseFilterSet):
             Q(name__icontains=value)
             | Q(url__icontains=value)
             | Q(api_key__icontains=value)
+        )
+
+
+class ObjectChangeFilterSet(BaseFilterSet):
+    q = django_filters.CharFilter(method="search", label="Search")
+    time = django_filters.DateTimeFromToRangeFilter()
+    action = django_filters.MultipleChoiceFilter(
+        choices=ObjectChangeAction, null_value=None
+    )
+    user_id = django_filters.ModelMultipleChoiceFilter(
+        queryset=User.objects.all(), label="User (ID)"
+    )
+    user = django_filters.ModelMultipleChoiceFilter(
+        field_name="user__username",
+        queryset=User.objects.all(),
+        to_field_name="username",
+        label="User name",
+    )
+
+    class Meta:
+        model = ObjectChange
+        fields = [
+            "id",
+            "user",
+            "user_name",
+            "request_id",
+            "action",
+            "changed_object_type_id",
+            "changed_object_id",
+            "object_repr",
+        ]
+
+    def search(self, queryset, name, value):
+        if not value.strip():
+            return queryset
+        return queryset.filter(
+            Q(user_name__icontains=value) | Q(object_repr__icontains=value)
         )
