+++ extras/forms.py
@@ -1,4 +1,5 @@
 from django import forms
+from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
 from requests.exceptions import HTTPError
@@ -9,12 +10,20 @@ from utils.forms.fields import (
     CommentField,
     ContentTypeChoiceField,
     DynamicModelChoiceField,
+    DynamicModelMultipleChoiceField,
     JSONField,
     SlugField,
 )
-from utils.forms.widgets import ColorSelect, CustomNullBooleanSelect, StaticSelect
+from utils.forms.widgets import (
+    APISelectMultiple,
+    ColorSelect,
+    CustomNullBooleanSelect,
+    StaticSelect,
+    StaticSelectMultiple,
+)
 
-from .models import IXAPI, ConfigContext, ExportTemplate, Tag
+from .enums import ObjectChangeAction
+from .models import IXAPI, ConfigContext, ExportTemplate, ObjectChange, Tag
 from .utils import FeatureQuery
 
 
@@ -96,6 +105,31 @@ class IXAPIForm(BootstrapMixin, forms.ModelForm):
 class IXAPIFilterForm(BootstrapMixin, forms.Form):
     model = IXAPI
     q = forms.CharField(required=False, label="Search")
+
+
+class ObjectChangeFilterForm(BootstrapMixin, forms.Form):
+    model = ObjectChange
+    q = forms.CharField(required=False, label="Search")
+    time_after = forms.DateTimeField(
+        label="After",
+        required=False,
+        widget=forms.TextInput(attrs={"placeholder": "YYYY-MM-DD hh:mm:ss"}),
+    )
+    time_before = forms.DateTimeField(
+        label="Before",
+        required=False,
+        widget=forms.TextInput(attrs={"placeholder": "YYYY-MM-DD hh:mm:ss"}),
+    )
+    action = forms.ChoiceField(
+        required=False, choices=ObjectChangeAction, widget=StaticSelectMultiple
+    )
+    user_id = DynamicModelMultipleChoiceField(
+        queryset=User.objects.all(),
+        required=False,
+        display_field="username",
+        label="User",
+        widget=APISelectMultiple(api_url="/api/users/users/"),
+    )
 
 
 class TagBulkEditForm(BootstrapMixin, BulkEditForm):
