+++ extras/tables.py
@@ -1,4 +1,5 @@
 import django_tables2 as tables
+from django.conf import settings
 
 from utils.tables import (
     BaseTable,
@@ -14,9 +15,34 @@ from .models import (
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
+    ObjectChange,
     Tag,
     TaggedItem,
 )
+
+OBJECT_CHANGE_ACTION = """
+{% if record.action == "create" %}
+<span class="badge badge-success">Created</span>
+{% elif record.action == "update" %}
+<span class="badge badge-primary">Updated</span>
+{% elif record.action == "delete" %}
+<span class="badge badge-danger">Deleted</span>
+{% endif %}
+"""
+
+OBJECT_CHANGE_OBJECT = """
+{% if record.action != 3 and record.changed_object.get_absolute_url %}
+<a href="{{ record.changed_object.get_absolute_url }}">{{ record.object_repr }}</a>
+{% elif record.action != 3 and record.related_object.get_absolute_url %}
+<a href="{{ record.related_object.get_absolute_url }}">{{ record.object_repr }}</a>
+{% else %}
+{{ record.object_repr }}
+{% endif %}
+"""
+
+OBJECT_CHANGE_REQUEST_ID = """
+<a href="{% url 'extras:objectchange_list' %}?request_id={{ value }}">{{ value }}</a>
+"""
 
 
 class ConfigContextTable(BaseTable):
@@ -78,6 +104,37 @@ class IXAPITable(BaseTable):
         model = IXAPI
         fields = ("name", "url", "api_key", "actions")
         default_columns = ("name", "url", "api_key", "actions")
+
+
+class ObjectChangeTable(BaseTable):
+    time = tables.DateTimeColumn(linkify=True, format=settings.SHORT_DATETIME_FORMAT)
+    action = tables.TemplateColumn(template_code=OBJECT_CHANGE_ACTION)
+    changed_object_type = tables.Column(verbose_name="Type")
+    object_repr = tables.TemplateColumn(
+        template_code=OBJECT_CHANGE_OBJECT, verbose_name="Object"
+    )
+    request_id = tables.TemplateColumn(
+        template_code=OBJECT_CHANGE_REQUEST_ID, verbose_name="Request ID"
+    )
+
+    class Meta(BaseTable.Meta):
+        model = ObjectChange
+        fields = (
+            "time",
+            "user_name",
+            "action",
+            "changed_object_type",
+            "object_repr",
+            "request_id",
+        )
+        default_columns = (
+            "time",
+            "user_name",
+            "action",
+            "changed_object_type",
+            "object_repr",
+            "request_id",
+        )
 
 
 class TagTable(BaseTable):
