+++ extras/tests/test_views.py
@@ -1,8 +1,11 @@
+import uuid
 from unittest.mock import patch
 
+from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 
-from extras.models import IXAPI, ConfigContext, ExportTemplate, Tag
+from extras.enums import ObjectChangeAction
+from extras.models import IXAPI, ConfigContext, ExportTemplate, ObjectChange, Tag
 from peering.models import AutonomousSystem
 from utils.testing import ViewTestCases
 
@@ -143,6 +146,30 @@ class IXAPITestCase(ViewTestCases.PrimaryObjectViewTestCase):
             ],
         ):
             super().test_edit_object_with_permission()
+
+
+class ObjectChangeTestCase(ViewTestCases.ReadOnlyObjectViewTestCase):
+    model = ObjectChange
+
+    test_changelog_object = None
+    test_create_object = None
+    test_edit_object = None
+    test_delete_object = None
+    test_bulk_edit_objects = None
+    test_bulk_delete_objects = None
+
+    @classmethod
+    def setUpTestData(cls):
+        tag = Tag(name="Tag 1", slug="tag-1")
+        tag.save()
+
+        user = User.objects.create_user(username="testuser2")
+        for i in range(1, 4):
+            uid = uuid.uuid4()
+            change = tag.to_objectchange(ObjectChangeAction.UPDATE)
+            change.user = user
+            change.request_id = uid
+            change.save()
 
 
 class TagTestCase(ViewTestCases.OrganizationalObjectViewTestCase):
