+++ extras/webhooks.py
@@ -5,9 +5,9 @@ from django.contrib.contenttypes.models import ContentType
 from django.utils import timezone
 from django_rq import get_queue
 
+from extras.enums import ObjectChangeAction
 from extras.models import Webhook
 from utils.api import get_serializer_for_model
-from utils.enums import ObjectChangeAction
 from utils.functions import serialize_object
