+++ templates/utils/object_change/details.html
@@ -1,7 +1,7 @@
 {% extends '_base.html' %}
 {% load helpers %}
 {% block breadcrumb %}
-<li class="breadcrumb-item"><a href="{% url 'utils:objectchange_list' %}">Change Log</a></li>
+<li class="breadcrumb-item"><a href="{% url 'extras:objectchange_list' %}">Change Log</a></li>
 {% if instance.related_object.get_absolute_url %}
 <li class="breadcrumb-item"><a href="{{ instance.related_object.get_absolute_url }}changelog/">{{ instance.related_object }}</a></li>
 {% elif instance.changed_object.get_absolute_url %}
@@ -55,10 +55,10 @@
             <div class="card-header">
               <strong>Difference</strong>
               <div class="btn-group float-right" role="group" aria-label="Change Navigation">
-                <a {% if previous_change %}href="{% url 'utils:objectchange_view' pk=previous_change.pk %}"{% endif %} class="btn btn-xs btn-secondary{% if not previous_change %} disabled{% endif %}">
+                <a {% if previous_change %}href="{% url 'extras:objectchange_view' pk=previous_change.pk %}"{% endif %} class="btn btn-xs btn-secondary{% if not previous_change %} disabled{% endif %}">
                   <i class="fas fa-chevron-left"></i> Previous
                 </a>
-                <a {% if next_change %}href="{% url 'utils:objectchange_view' pk=next_change.pk %}"{% endif %} class="btn btn-xs btn-secondary{% if not next_change %} disabled{% endif %}">
+                <a {% if next_change %}href="{% url 'extras:objectchange_view' pk=next_change.pk %}"{% endif %} class="btn btn-xs btn-secondary{% if not next_change %} disabled{% endif %}">
                   Next <i class="fas fa-chevron-right"></i>
                 </a>
               </div>
@@ -91,7 +91,7 @@
                 {% endspaceless %}
 {% endfor %}</pre>
               {% elif non_atomic_change %}
-              Warning: Comparing non-atomic change to previous change record (<a href="{% url 'utils:objectchange_view' pk=previous_change.pk %}">{{ previous_change.pk }}</a>)
+              Warning: Comparing non-atomic change to previous change record (<a href="{% url 'extras:objectchange_view' pk=previous_change.pk %}">{{ previous_change.pk }}</a>)
               {% else %}
               <span class="text-muted">None</span>
               {% endif %}
@@ -119,7 +119,7 @@
           {% include 'utils/card_table.html' with table=related_changes_table header='Related Changes' %}
           {% if related_changes_count > related_changes_table.rows|length %}
           <div class="float-right mt-2">
-            <a href="{% url 'utils:objectchange_list' %}?request_id={{ instance.request_id }}" class="btn btn-primary">See all {{ related_changes_count | add:"1" }} changes</a>
+            <a href="{% url 'extras:objectchange_list' %}?request_id={{ instance.request_id }}" class="btn btn-primary">See all {{ related_changes_count | add:"1" }} changes</a>
           </div>
           {% endif %}
         </div>
