+++ utils/enums.py
@@ -102,15 +102,3 @@ class Colour(ChoiceSet):
         (BLACK, "Black"),
         (WHITE, "White"),
     )
-
-
-class ObjectChangeAction(ChoiceSet):
-    CREATE = "create"
-    UPDATE = "update"
-    DELETE = "delete"
-
-    CHOICES = (
-        (CREATE, "Created", "success"),
-        (UPDATE, "Updated", "warning"),
-        (DELETE, "Deleted", "danger"),
-    )
