+++ utils/filters.py
@@ -3,14 +3,11 @@ from copy import deepcopy
 import django_filters
 from django import forms
 from django.conf import settings
-from django.contrib.auth.models import User
 from django.db import models
 from django.db.models import Q
 from django_filters.constants import EMPTY_VALUES
 
-from utils.enums import ObjectChangeAction
 from utils.forms.fields import multivalue_field_factory
-from utils.models import ObjectChange
 
 
 class ContentTypeFilter(django_filters.CharFilter):
@@ -154,40 +151,3 @@ class NameSlugSearchFilterSet(django_filters.FilterSet):
         if not value.strip():
             return queryset
         return queryset.filter(Q(name__icontains=value) | Q(slug__icontains=value))
-
-
-class ObjectChangeFilterSet(BaseFilterSet):
-    q = django_filters.CharFilter(method="search", label="Search")
-    time = django_filters.DateTimeFromToRangeFilter()
-    action = django_filters.MultipleChoiceFilter(
-        choices=ObjectChangeAction, null_value=None
-    )
-    user_id = django_filters.ModelMultipleChoiceFilter(
-        queryset=User.objects.all(), label="User (ID)"
-    )
-    user = django_filters.ModelMultipleChoiceFilter(
-        field_name="user__username",
-        queryset=User.objects.all(),
-        to_field_name="username",
-        label="User name",
-    )
-
-    class Meta:
-        model = ObjectChange
-        fields = [
-            "id",
-            "user",
-            "user_name",
-            "request_id",
-            "action",
-            "changed_object_type_id",
-            "changed_object_id",
-            "object_repr",
-        ]
-
-    def search(self, queryset, name, value):
-        if not value.strip():
-            return queryset
-        return queryset.filter(
-            Q(user_name__icontains=value) | Q(object_repr__icontains=value)
-        )
