+++ utils/forms/__init__.py
@@ -1,13 +1,8 @@
 from django import forms
-from django.contrib.auth.models import User
 from django.db.models import Count
 from taggit.forms import TagField
 
-from utils.enums import ObjectChangeAction
-from utils.models import ObjectChange
-
-from .fields import DynamicModelMultipleChoiceField
-from .widgets import APISelectMultiple, StaticSelectMultiple
+from .widgets import StaticSelectMultiple
 
 
 def add_blank_choice(choices):
@@ -87,31 +82,6 @@ class TableConfigurationForm(BootstrapMixin, forms.Form):
     @property
     def table_name(self):
         return self.table.__class__.__name__
-
-
-class ObjectChangeFilterForm(BootstrapMixin, forms.Form):
-    model = ObjectChange
-    q = forms.CharField(required=False, label="Search")
-    time_after = forms.DateTimeField(
-        label="After",
-        required=False,
-        widget=forms.TextInput(attrs={"placeholder": "YYYY-MM-DD hh:mm:ss"}),
-    )
-    time_before = forms.DateTimeField(
-        label="Before",
-        required=False,
-        widget=forms.TextInput(attrs={"placeholder": "YYYY-MM-DD hh:mm:ss"}),
-    )
-    action = forms.ChoiceField(
-        required=False, choices=ObjectChangeAction, widget=StaticSelectMultiple
-    )
-    user_id = DynamicModelMultipleChoiceField(
-        queryset=User.objects.all(),
-        required=False,
-        display_field="username",
-        label="User",
-        widget=APISelectMultiple(api_url="/api/users/users/"),
-    )
 
 
 class AddRemoveTagsForm(forms.Form):
